--- src/components/Icon/types.ts
@@ -1,9 +1,0 @@
-/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export enum IconSize {
-  XXLARGE = '32',
-  XLARGE = '24',
-  LARGE = '20',
-  MEDIUM = '16',
-  SMALL = '14',
-  XSMALL = '12',
-}
