+++ src/components/Illustration/BaseIllustration.tsx
@@ -2,23 +2,8 @@
 import '../../yamui';
 import * as React from 'react';
 import { join } from '../../util/classNames';
-import { BaseComponentProps } from '../../util/BaseComponent/props';
 import { getClassNames } from './Illustration.styles';
-
-export interface IllustrationProps extends BaseComponentProps {
-  /**
-   * Sets illustration style to `display: block`. This is helpful if the inline
-   * illustration's `line-height` is taller than the current line's
-   * `line-height`, and is creating extra vertical space.
-   */
-  block?: boolean;
-
-  /**
-   *  If provided, this pixel value will override the default size this Illustration is optimized for.
-   */
-  size?: number;
-}
-
+import { IllustrationProps } from './Illustration.types';
 /**
  * An `Illustration` renders an SVG illustration. Illustrations come in multiple
  * sizes, so you must import the proper size for your application.
