+++ src/components/Illustration/index.ts
@@ -1,2 +1,3 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export { default as BaseIllustration, IllustrationProps } from './BaseIllustration';
+export { default as BaseIllustration } from './BaseIllustration';
+export * from './Illustration.types';
