--- src/components/Image/types.ts
@@ -1,5 +1,0 @@
-/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export enum BorderType {
-  SOFT = 'soft',
-  ROUND = 'round',
-}
